Move the codegen package into the compiler package.
